app/requirements/dev.txt
gunicorn>=21.2.0
django-environ>=0.10.0
django-debug-toolbar>=4.1.0
Django>=4.2.4
django-storages>=1.13.2
django-widget-tweaks>=1.4.12
google-api-python-client==2.95.0
setuptools>=65.5.1
psycopg2==2.9.7
celery==5.3.1
asyncio==3.4.3
isodate==0.6.1
openai==0.27.8
python-dotenv==1.0.0
python-dateutil==2.8.2
pytube==15.0.0
spotipy==2.23.0
boto3>=1.28.21
djangorestframework>=3.14.0
django-rest-swagger>=2.2.0
-U drf-yasg